+++ src/euroeval/cli.py
@@ -235,6 +235,22 @@ from .languages import get_all_languages
     help="""This option is deprecated - please use --language instead.""",
 )
 @click.option(
+    "--max-context-length",
+    default=None,
+    type=int,
+    show_default=True,
+    help="Override for the maximum context length of the model. If not specified, the "
+    "value will be inferred automatically from the model.",
+)
+@click.option(
+    "--vocabulary-size",
+    default=None,
+    type=int,
+    show_default=True,
+    help="Override for the vocabulary size of the model. If not specified, the value "
+    "will be inferred automatically from the model.",
+)
+@click.option(
     "--batch-size",
     default=None,
     type=click.Choice(["1", "2", "4", "8", "16", "32"]),
@@ -271,6 +287,8 @@ def benchmark(
     debug: bool,
     model_language: tuple[str],
     dataset_language: tuple[str],
+    max_context_length: int | None,
+    vocabulary_size: int | None,
     batch_size: str | None,
 ) -> None:
     """Benchmark pretrained language models on language tasks."""
@@ -307,6 +325,8 @@ def benchmark(
         model_language=None if len(model_language) == 0 else list(model_language),
         dataset_language=None if len(dataset_language) == 0 else list(dataset_language),
         batch_size=int(batch_size) if batch_size is not None else None,
+        max_context_length=max_context_length,
+        vocabulary_size=vocabulary_size,
     ).benchmark(model=list(model))