+++ tests/test_cli.py
@@ -36,5 +36,7 @@ def test_cli_param_names(cli_params: dict[str, ParamType]) -> None:
         "model_language",
         "dataset_language",
         "batch_size",
+        "max_context_length",
+        "vocabulary_size",
         "help",
     }