+++ CHANGELOG.md
@@ -9,8 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- Added the new Danish linguistic acceptability dataset DaLA. It's marked as 
+- Added the new Danish linguistic acceptability dataset DaLA. It's marked as
   unofficial for now. This was added by @N-essuno ✨
+- Added `--max-context-length` and `--vocabulary-size` CLI options (and corresponding
+  `max_context_length` and `vocabulary_size` arguments to `Benchmarker.__init__` and
+  `Benchmarker.benchmark`) to allow overriding the model metadata values that are
+  inferred automatically from the model. This is useful when the model does not have
+  the metadata specified, or has it specified incorrectly.
 
 ## [v16.16.1] - 2026-02-25